requirements.txt
flask
flask-cors
rembg
pillow
opencv-python-headless
numpy
gunicorn